The new model skeleton approach is now implemented.
Basically the model system is working as before but with a completely revised and simplified dataset class, and major changes in the models.py class as well.  Most of the functionality that appears to be standard I moved into skeleton.py which gives a skeleton set of urbansim models to start with.  Still not sure whether these need to be in urbansim or in region specific implementations.