changes to update to the latest v3 data and zoning data
This has a few changes to work towards getting simulations running again.

New hedonics are estimated, using the RedFin data that Eddie included.  It looks pretty good so far.

The new zoning is also integrated and zoning well to the parcel data.  I had to make a few changes to integrate this data, but nothing major.