add comments/links to urbansim.org documentation of data
hopefully the documentation linked to on urbansim.org is informative. perhaps we find a better spot for this in the future but for now this might help people looking to understand the data in the h5 files. also, had to resolve merge conflict where line on zoning_for_parcels table had been removed since this original (cherry-picked) commit was made.