simplifying geom aggregation to speed it up
by changing a handful of lines, this script is more honest about what it is: SQL. so now we can do things like vacuum (analyze), which is known to improve performance after many of the operations performed in this script (updating many thousand of rows, for example) without having to do python configuration. this also makes it easier to understand, since now its only 1 language. the only remaining issue is to either re-write the python function called from spatialtoolz in spandex as SQL, or else break this script into components and leave that python at the middle of those components.