remove unused columns in zoninglookup-add plandate
removed unused columns from the zoning_lookup table. we can always add them back later using this history and the mapping between the objectid and the id column added for 2006 data.
the 2006 table contains a plandate column, and it seems useful so we will add it to the generic table.