change 2006 table to insert into zoning_lookup
in the 2006 data 'juris' was a text field which is 'city' in the zoning_lookup table. renamed accordingly

there was not an entry for 'juris' as an integer as in the zoning_lookup table. left it empty.

the 'origgplu' column seems to be the equiavalent of the 'name' column in the zoning_lookup table. renamed accordingly.

deleted various columns that were not in the table. moved plandate to the last column to match zoning_lookup