match format of 2006 and 2012 0's, 1's and x's buildable type indicator `blank/x` to `0/1`
in the source data buildable types are given an `x` if true. here we change this to 0 if false, 1 if true. previously, we did this in the zoning processing steps. but doing it here makes more sense for python/bayarea_urbansim usage. we can just remove the lines from the database piece