add zoning/parcel lookup tables
adding the table(s) that say which zoning_id a parcel has.
the qualities of this zoning id can be found in zoning_lookup.csv.
i added two versions of these tables.
in zoning_parcels.csv the zoning_id that was assigned to the parcel is kept,
regardless of whether it is a nodev parcel.
in the zoning_parcels_nodev_remove_zoning_id.csv the zoning_id is changed to 99999,
which has a corresponding row in the zoning_lookup.csv table.
@fscottfoti, please let me know which of these you will need for future development.
the former is preferable from a standpoint of not losing information but i understand you may have specific needs.

the relevant commit from the zoning repo is this: https://github.com/MetropolitanTransportationCommission/zoning/commit/809bc3cda4dd1dd1f0dca6afa38335c67458ca85