update readme with data links and instructions
simulation.py runs through to completion with this set of data on master. in the future, we should always check that simulation.py completes without exception with the data that are linked to above and through make.

readme:database_backup/parcel_shapefile targets

adding notes on how to get the database backup for a particular h5. this should help with debugging data issues in the future.