read unstacked control-totals into legacy table
the old control totals data was stacked in the sense that, for example, the value in the sum_of_households column was given for a particular income quartile, where the income quartile was a value in the income_quartile column. the new controls totals data contains a column for each income quartile. so here we take the new data, stack it by column names, and then fill in the columns and rename them to match the format of the old data frame. we do this for both employment_controls and household_controls.