temporary patch to bart3a/3b issue in ttp_upzone
in the tpp upzone file, we had an inconsistency between the shapefile
which described upzoned regions and the table called zoning_mods. here we
remove the bart3a column to resolve this temporarily. we do this because
there are major changes and improvements in the data otherwise, and since
this bug is also in the previous tpp_upzone file.