add base year (aka 2009) data to git
we should version these data. they are not large in terms of file size, and they are basically just summaries of whats in the data directory. i did a preliminary comparison of them across runs and they do not exhibit any differences at the level of specificity that we use to report them in the output assesment.