cleanup, formatting, pep8
this commit cleans up the changes to the previous one. in the this one and the previous one, fix the way ratios and the change in the share of the variable are calculated. previously, ratios related the run outcome to the base year variable. in this one they are used to compare across runs.