scale the output acres to match inputs from abag
abag has different acres numbers in the base year for residential and commercial and industrial properties than what we have in the data we are using. so here we take the raio of change output by the simulation model we are using, and the multiply their base year numbers by that.