write commercial.residential taz acre csv with max(totacre) conditional
in 300 taz's, for commercial and industrial acres, a linear model of change in acres based on urbansim inputs and outputs put the acres for this type over the totalacres for the taz. and in 100 taz's the same was the case for resacre. so here we limit the amount for those variables in each taz to 2x the total acres in the taz.