set default geography to taz in output_csv_utils
now that geography is a global variable in output_csv_utils, the functions
in it have had that argument removed from them. but in fix_acres we call
specific functions that used to have that argument. so here we set the default global variable for geography to taz to make fix_acres work, because the compare_outcome script sets the global variable for geography each time it calls the compare function.