allow user to enter non-numerically order sim runs to compare output
occasionally the user will want to compare outputs that are not in numeric order of the run number. for example, if scenario number 2 (or 'pr'), is actually a run number that is lower in number than scenario number 1, then this might happen. this commit changes the script to allow for that.