logic to create subsidized units
This takes the amount of subsidy for a development and divides by the expected revenue per unit in order to get the number of units that are deed restricted for a given development.  Partial units ARE preserved, so if you had 4 "quarter units", the 4th development considered would get 1 subsidized unit.