link to parcel geometries; describe geomid
following the lead of la county
(http://egis3.lacounty.gov/dataportal/2015/09/11/california-statewide-parcel-boundaries/), it seems that there is no restriction on publishing parcel bounaries. here we have removed the assessed value column from the parcels file output by the dsata regenration process and posted a link to it. in addition, based on feedback from bobbylu, we have added a short description of what the geom_id is.