adding more generalized acct system
I realized 3 policies we work with in the bay area take the same form - these are the parcel tax, capital gains, and OBAG - they take a lump sum regional payment and distribute it to a subset of parcels.  These I'm calling lump sum policies and can be configured directly in yaml.

Second the CEQA and parking requirements are considered pct of profitability adjustments and these can be specified in YAML as well.