add ability to downzone and drop buildings
this works completely analagously to the scenario-based upzoning and adding of buildings types

One oddity is that downzing overrides upzoning.  keep in mind the residential building types are equal to each other - there are in effect "uses" where the density comes from the dua or far.  What this means is that if you remove HS as a building type, you might think you are removing low density residential only - NO!! - the density comes from the dua and far attributes.  *If you drop any residential buildings type you will effectively be dropping residential!*

(In the future I hope to use the building types in the developer model quite a bit more than the densities, perhaps just using the densities to confirm the allowable building types, or to give a specific density to a building type.  This is IN THE FUTURE.)