actually override individual values in the development limits
rather than overriding ALL of defaults with all of a scenario's limits, do it item by item.  A limit for a certain city in the default will still apply in the scenario unless it is overriden with a key that is the same in the scenario as in the default.