use a region-wide proportion of jobs to households
rather than one that varies per jurisdiction.  We expect there to be a higher ratio of local serving retail in Cloverdale e.g. than San Francisco, so we take the ratio of local serving retail to be 1/3rd of the regional total of retail divided by the number of households.  As Cloverdale grows (if it does), we expect the local serving retail to increase according to that ratio, no ifs ands and buts, and we wouldn't expect the ELCM to get this logic right necessarily.