creating data preprocessing step
many major preprocessing steps were performed when tables were accessed. this is a major benefit of orca - you can write transformations to data tables in code and perform them every time the table is accessed, and share the code via github.

there are limits to this, which mainly have to do with processing time.  for longer running processes, it's not efficient to run every time a table is accessed in order to do interactive exploratory work. there's no hard and fast rule for what crosses, the line, but for deed restricted units and adjustments of vacancy rates, we had probably crossed it.

there is now a "preprocessing" MODE in run.py which runs the steps which currently fall under this rubric, as well as a preprocessing.py file where the code lives.  when you get the .h5 file from s3, you will need to run this mode before running anything else.

as part of this change, run.py now has a proper usage statement using argparse to get arguments.  the "--mode preprocessing" argument is used to run preprocessing.