bug fixes to get new base year summaries to work
we now do a base year summary every run (rather than having a baseyearsim mode.  in the first simulation year, we just write out the summaries and from then one we run the full set of models.  the "-k" option of run.py is used to skip the base year when debugging.